Support running tests with unittest module
For tests to be discovered, they have to be in a proper package; the
'tests' folder makes this not a proper package (it has no __init__
file), instead we can just keep it in the root folder for the module.

Use relative imports when importing modules relative to a particular
path. Also fixes up an issue starting the test.