Don't import version
Yeah... that's not a good idea. It has to traverse the module chain, and
when it hits the __init__.py for the top-level module, it loads all
kinds of stuff (b/c that file has a lot of stuff exposed), which loads
dependencies, which aren't available yet if you're running an install...
blah.