Replace default filebrowser
The default filebrowser used in JupyterLab is hard to replace, yet it's
possible. The way this is done is by manually disabling the default
plugin ("filebrowser-extensions:factory") and then essentially
replicating what that plugin does (creates the file browser) in another
plugin that you control.

As an initial test, we're just extending the default renderer to add
some additional (invisible) UI to the list items.

This code WILL be annoying to maintain over time, I'd wager. We have to
ensure that the factory setup doesn't add anything major in the default
factory, since any code changes there would not be picked up here and
would have to be manually copied.