Store directories as relative to notebook_dir
This makes a lot of things easier, especially when trying to use this
data in the client-side code: we can reference the same relative paths
that JupyterLab's contents APIs use. Also removes some of the complexity
around storing the filepath... we initially talked about this being nice
but I think we can safely give it a generic filename and that makes it
easier to deal with the create vs. update use case as well.