Separate upload and update redirect URLs
This breaks up some of the shared code between the two handlers and opts
for duplicating a bit of the functionality that happens on success. Also
makes a few changes like exposing the config object directly on the
handler to make it easier to retrieve more values in the future.